deposit transfer to spec
