+++ logicrunner/builtin/proxy/wallet/wallet.go
@@ -463,105 +463,6 @@ func (r *Wallet) GetBalanceAsImmutable(assetName string) (string, error) {
 	return ret0, nil
 }
 
-// Accept is proxy generated method
-func (r *Wallet) Accept(amountStr string, assetName string) error {
-	var args [2]interface{}
-	args[0] = amountStr
-	args[1] = assetName
-
-	var argsSerialized []byte
-
-	ret := make([]interface{}, 1)
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "Accept", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	resultContainer := foundation.Result{
-		Returns: ret,
-	}
-	err = common.CurrentProxyCtx.Deserialize(res, &resultContainer)
-	if err != nil {
-		return err
-	}
-	if resultContainer.Error != nil {
-		err = resultContainer.Error
-		return err
-	}
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// AcceptNoWait is proxy generated method
-func (r *Wallet) AcceptNoWait(amountStr string, assetName string) error {
-	var args [2]interface{}
-	args[0] = amountStr
-	args[1] = assetName
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "Accept", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// AcceptAsImmutable is proxy generated method
-func (r *Wallet) AcceptAsImmutable(amountStr string, assetName string) error {
-	var args [2]interface{}
-	args[0] = amountStr
-	args[1] = assetName
-
-	var argsSerialized []byte
-
-	ret := make([]interface{}, 1)
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "Accept", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	resultContainer := foundation.Result{
-		Returns: ret,
-	}
-	err = common.CurrentProxyCtx.Deserialize(res, &resultContainer)
-	if err != nil {
-		return err
-	}
-	if resultContainer.Error != nil {
-		err = resultContainer.Error
-		return err
-	}
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
 // AddDeposit is proxy generated method
 func (r *Wallet) AddDeposit(txId string, deposit insolar.Reference) error {
 	var args [2]interface{}
